infra/requirements.txt
aws-cdk-lib>=2.100.0 # Using a recent version, adjust as needed
constructs>=10.0.0 